fix(windows): 修复 HANDLE 类型不匹配导致 CI 构建失败
windows-sys 0.59 的 HANDLE 为 *mut c_void，
去掉多余的 as isize 转换。

保留 cargo 预构建调试步骤验证修复。

+++ native/hub/src/segment_coordinator.rs
@@ -520,7 +520,7 @@ pub async fn run_coordinated_download(
                     struct FileAllocInfo {
                         allocation_size: i64,
                     }
-                    let handle = std_file.as_raw_handle() as isize;
+                    let handle = std_file.as_raw_handle();
                     // Step 1: 预分配 NTFS 物理簇——立即保留磁盘空间（连续簇优先），
                     // 磁盘不足时提前报错（等效 Linux fallocate 的 ENOSPC 检测），
                     // 减少多段随机写时的 NTFS 碎片化。
